many changes

+++ client/src/lib/data/datamodel.svelte.ts
@@ -22,18 +22,64 @@ export abstract class Repo<
 
 	public all = $derived(Object.values(this.store));
 
-	// Protected hooks for subclasses to override
-	protected abstract get basePath(): string;
+	// Static helpers that concrete repos inherit
+	static get idParam(): string {
+		const p = (this as any).path as string;
+		if (!p) throw new Error('Missing static path on Repo subclass');
+		return inferIdParam(p);
+	}
+
+	static async remoteList<ListParams = unknown, T = unknown>(params?: ListParams): Promise<T[]> {
+		const { api } = await import('../api/client');
+		const res = await api.GET((this as any).path as any, { params: { query: (params as any) ?? {} } as any });
+		return (res.data as T[]) ?? [];
+	}
+
+	static async remoteGet<T = unknown>(id: Id): Promise<T> {
+		const { api } = await import('../api/client');
+		const itemPath = `${(this as any).path}/{${(this as any).idParam}}`;
+		const res = await api.GET(itemPath as any, { params: { path: { [(this as any).idParam]: Number(id) } } as any });
+		if (!res.data) throw new Error('No data');
+		return res.data as T;
+	}
+
+	static async remoteCreate<T = unknown, B = unknown>(body: B): Promise<T> {
+		const { api } = await import('../api/client');
+		const res = await api.POST((this as any).path as any, { body } as any);
+		if (!res.data) throw new Error('No data');
+		return res.data as T;
+	}
+
+	static async remoteUpdate<T = unknown, B = unknown>(id: Id, body: B): Promise<T> {
+		const { api } = await import('../api/client');
+		const itemPath = `${(this as any).path}/{${(this as any).idParam}}`;
+		const res = await api.PATCH(itemPath as any, { params: { path: { [(this as any).idParam]: Number(id) } } as any, body } as any);
+		if (!res.data) throw new Error('No data');
+		return res.data as T;
+	}
+
+	static async remoteDelete(id: Id): Promise<void> {
+		const { api } = await import('../api/client');
+		const itemPath = `${(this as any).path}/{${(this as any).idParam}}`;
+		await api.DELETE(itemPath as any, { params: { path: { [(this as any).idParam]: Number(id) } } as any });
+	}
+
+	// Protected hooks for subclasses to override or inherit
+	protected get basePath(): string {
+		return (this.constructor as any).path as string;
+	}
 
 	protected get idParam(): string {
 		return inferIdParam(this.basePath);
 	}
 
-	// Optional object factory for subclasses
-	protected createDataObject?(id: Id): TDataObject;
+	// Optional object factory for subclasses (now takes a row)
+	protected createDataObject?(obj: TGet): TDataObject;
 
-	public object(id: Id): TDataObject {
-		return this.createDataObject ? this.createDataObject(id) : (new DataObject<TGet, TPatch>(this, id) as TDataObject);
+	public object(objOrId: Id): TDataObject {
+		const found = this.get(objOrId as Id);
+		if (!found) throw new Error(`Missing row ${String(objOrId)}`);
+		return this.createDataObject ? this.createDataObject(found) : (new DataObject<TGet, TPatch>(found, this) as TDataObject);
 	}
 
 	// Optional serialization hooks
@@ -45,49 +91,25 @@ export abstract class Repo<
 		return body;
 	}
 
-	// Remote operations - can be overridden by subclasses, all return TGet
+	// Remote operations - delegate to static versions
 	protected async remoteList(params?: ListParams): Promise<TGet[]> {
-		const { api } = await import('../api/client');
-		const res = await api.GET(this.basePath as any);
-		return (res.data as TGet[]) ?? [];
+		return (this.constructor as typeof Repo).remoteList<ListParams, TGet>(params as any);
 	}
 
 	protected async remoteGet(id: Id): Promise<TGet> {
-		const { api } = await import('../api/client');
-		const itemPath = `${this.basePath}/{${this.idParam}}`;
-		const res = await api.GET(itemPath as any, { 
-			params: { path: { [this.idParam]: Number(id) } } as any 
-		});
-		if (!res.data) throw new Error('No data');
-		return res.data as TGet;
+		return (this.constructor as typeof Repo).remoteGet<TGet>(id);
 	}
 
 	protected async remoteCreate(body: TCreate): Promise<TGet> {
-		const { api } = await import('../api/client');
-		const res = await api.POST(this.basePath as any, { 
-			body: this.toCreateBody(body) 
-		} as any);
-		if (!res.data) throw new Error('No data');
-		return res.data as TGet;
+		return (this.constructor as typeof Repo).remoteCreate<TGet, TCreate>(this.toCreateBody(body) as any);
 	}
 
 	protected async remoteUpdate(id: Id, body: TPatch): Promise<TGet> {
-		const { api } = await import('../api/client');
-		const itemPath = `${this.basePath}/{${this.idParam}}`;
-		const res = await api.PATCH(itemPath as any, { 
-			params: { path: { [this.idParam]: Number(id) } } as any, 
-			body: this.toUpdateBody(body) 
-		} as any);
-		if (!res.data) throw new Error('No data');
-		return res.data as TGet;
+		return (this.constructor as typeof Repo).remoteUpdate<TGet, TPatch>(id, this.toUpdateBody(body) as any);
 	}
 
 	protected async remoteDelete(id: Id): Promise<void> {
-		const { api } = await import('../api/client');
-		const itemPath = `${this.basePath}/{${this.idParam}}`;
-		await api.DELETE(itemPath as any, { 
-			params: { path: { [this.idParam]: Number(id) } } as any 
-		});
+		return (this.constructor as typeof Repo).remoteDelete(id);
 	}
 
 	get(id: Id): TGet | undefined {
@@ -121,60 +143,98 @@ export abstract class Repo<
 	async fetchOne(id: Id): Promise<DataObject<TGet, TPatch>> {
 		const row = await this.remoteGet(id);
 		this.upsert(row);
-		return this.object(id);
+		return this.object(row);
 	}
 
 	async create(body: TCreate): Promise<DataObject<TGet, TPatch>> {
 		const created = await this.remoteCreate(body);
 		this.upsert(created);
-		return this.object(created.id);
+		return this.object(created);
 	}
 }
 
 export class DataObject<TGet extends { id: Id }, TPatch = Partial<TGet>> {
 	// Make data reactive using $derived
 	data: TGet | undefined;
+	$: TGet
+	protected repo: Repo<TGet, any, TPatch, any, any> | undefined;
+	public readonly id: Id;
+	private _local: TGet | undefined = $state(undefined);
 
-	constructor(protected repo: Repo<TGet, any, TPatch, any, any>, public readonly id: Id) {
-		this.data = $derived(this.repo.store[this.id] as TGet | undefined);
+	constructor( obj: TGet, _repo?: Repo<TGet, any, TPatch, any, any>) {
+		this.repo = _repo;
+		this.id = obj.id;
+		if (!this.repo) {
+			this._local = obj;
+		}
+		this.data = $derived(this.repo ? this.repo.store[this.id] : this._local);
+		this.$ = $derived(this.data as TGet);
 	}
 
-	get required(): TGet {
-		const d = this.data;
-		if (!d) throw new Error(`Missing row ${String(this.id)}`);
-		return d;
-	}
+	// get $(): TGet {
+	// 	const d = this.data;
+	// 	if (!d) throw new Error(`Missing row ${String(this.id)}`);
+	// 	return d;
+	// }
 
 	updateLocal(patch: Partial<TGet>): TGet | undefined {
-		const curr = this.repo.store[this.id];
-		if (!curr) return undefined;
-		const next = { ...curr, ...patch } as TGet;
-		this.repo.store[this.id] = next;
-		return next;
+		if (this.repo) {
+			const curr = this.repo.store[this.id];
+			if (!curr) return undefined;
+			const next = { ...curr, ...patch } as TGet;
+			this.repo.store[this.id] = next;
+			return next;
+		}
+		if (!this._local) return undefined;
+		this._local = { ...this._local, ...patch } as TGet;
+		return this._local;
 	}
 
 	replace(next: TGet): TGet {
-		this.repo.store[this.id] = next;
+		if (this.repo) {
+			this.repo.store[this.id] = next;
+			return next;
+		}
+		this._local = next;
 		return next;
 	}
 
 	deleteLocal(): void {
-		this.repo.deleteLocal(this.id);
+		if (this.repo) {
+			this.repo.deleteLocal(this.id);
+			return;
+		}
+		this._local = undefined;
 	}
 
 	async refresh(): Promise<TGet> {
-		const obj = await this.repo.fetchOne(this.id);
-		return (obj as DataObject<TGet, TPatch>).required;
+		if (this.repo) return (await this.repo.fetchOne(this.id)).$ as TGet;
+		const RepoClass = (this.constructor as any).DefaultRepo as typeof Repo;
+		const next = await RepoClass.remoteGet<TGet>(this.id);
+		this._local = next;
+		return next;
 	}
 
 	async save(patch?: TPatch): Promise<TGet> {
-		const updated = await (this.repo as any).remoteUpdate(this.id, (patch ?? (this.required as unknown as TPatch)));
+		if (this.repo) {
+			const updated = await (this.repo as any).remoteUpdate(this.id, (patch ?? (this.$ as unknown as TPatch)));
+			this.replace(updated as TGet);
+			return updated as TGet;
+		}
+		const RepoClass = (this.constructor as any).DefaultRepo as typeof Repo;
+		const updated = await RepoClass.remoteUpdate<TGet, TPatch>(this.id, (patch ?? (this.$ as unknown as TPatch)));
 		this.replace(updated as TGet);
 		return updated as TGet;
 	}
 
 	async destroy(): Promise<void> {
-		await (this.repo as any).remoteDelete(this.id);
+		if (this.repo) {
+			await (this.repo as any).remoteDelete(this.id);
+			this.deleteLocal();
+			return;
+		}
+		const RepoClass = (this.constructor as any).DefaultRepo as typeof Repo;
+		await RepoClass.remoteDelete(this.id);
 		this.deleteLocal();
 	}
 }
